race condition if thread has 0 files
If `t.pruneOldestThreadFiles()` is aggressive and deletes all it's files, then the `for` loop will never end if the disk space still doesn't meet the threshold. Since the check for zero files is outside the for loop, it would never return. This needs to return if another thread also needs to delete files in the case threads are writing to the same disk.